requirements.txt
duckduckgo-search>=5.3.1
PyYAML>=6.0.1
